+++ src/main/java/com/solvd/atm/service/AtmService.java
@@ -3,6 +3,7 @@ package com.solvd.atm.service;
 import com.solvd.atm.domain.Account;
 import com.solvd.atm.domain.Atm;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 import java.util.List;
 
@@ -22,7 +23,7 @@ public interface AtmService {
      * operation select
      * quantity of many
      */
-    void getMoney(Account account, Integer money);
+    void getMoney(Account account, BigDecimal money);
 
     /**
      * finish work with current account
@@ -34,10 +35,10 @@ public interface AtmService {
 
     List<List<?>> moneyVariants(LinkedHashMap<Integer, Integer> cashInAtm, Integer requiredCash);
 
-    void transferMoney(Account account, String cardNumber, Integer money);
+    void transferMoney(Account account, String cardNumber, BigDecimal money);
 
     void continueWork();
 
-    boolean checkBalance(Integer money);
+    boolean checkBalance(BigDecimal money);
 
 }
